# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  domain.PaymentMethod:
    enum:
    - PIX
    - BOLETO
    type: string
    x-enum-varnames:
    - PaymentMethodPix
    - PaymentMethodBoleto
  domain.PaymentStatus:
    enum:
    - PAGAMENTO_PENDENTE
    - PAGO
    - REEMBOLSADO
    type: string
    x-enum-varnames:
    - PaymentPending
    - PaymentPaid
    - PaymentRefunded
...
  domain.Reservation:
    properties:
      cancellation_reason:
        type: string
      description:
        type: string
      end_time:
        type: string
      keys_returned_at:
        type: string
      keys_taken_at:
        type: string
      payment_amount:
        type: number
      payment_date:
        type: string
      payment_method:
        $ref: '#/definitions/domain.PaymentMethod'
      payment_status:
        $ref: '#/definitions/domain.PaymentStatus'
      resident_id:
        type: integer
      space:
        $ref: '#/definitions/domain.SpaceType'
      start_time:
        type: string
      status:
        $ref: '#/definitions/domain.ReservationStatus'
    type: object
  domain.ReservationStatus:
    enum:
    - PENDENTE
    - CONFIRMADA
    - CANCELADA
    - CHAVES_RETIRADAS
    - CHAVES_DEVOLVIDAS
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusConfirmed
    - StatusCancelled
    - StatusKeysTaken
    - StatusKeysReturned
...
  domain.SpaceType:
    enum:
    - SALAO_1
    - SALAO_2
    type: string
    x-enum-varnames:
    - Salon1
    - Salon2
...
host: localhost:8080
info:
  contact:
    email: support@portarius.com
    name: API Support
    url: http://www.portarius.com/support
  description: API for managing a package delivery system in a residential building
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Portarius API
  version: "1.0"
paths:
...
  /reservations:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of all reservations
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Reservation'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Get all reservations
      tags:
      - Reservations
    post:
      consumes:
      - application/json
      description: Creates a new reservation with the provided details
      parameters:
      - description: Reservation data
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/domain.Reservation'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a new reservation
      tags:
      - Reservations
  /reservations/{id}:
    delete:
      description: Delete a reservation by its ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete a reservation
      tags:
      - Reservations
    get:
      consumes:
      - application/json
      description: Retrieves a reservation based on its unique ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Get a reservation by ID
      tags:
      - Reservations
    put:
      consumes:
      - application/json
      description: Update reservation's time or description
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated data
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/domain.Reservation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update a reservation
      tags:
      - Reservations
  /reservations/{id}/cancel:
    put:
      description: Cancel a reservation by setting its status to cancelled and adding
        a cancellation reason
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cancellation body (only CancellationReason will be used)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Reservation'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Cancel a reservation
      tags:
      - Reservations
  /reservations/{id}/complete:
    put:
      description: Updates the reservation status to "keys returned" (complete)
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Mark reservation as complete
      tags:
      - Reservations
  /reservations/{id}/confirm:
    put:
      description: Update the reservation status to confirmed
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Confirm a reservation
      tags:
      - Reservations
  /reservations/{id}/confirm-payment:
    put:
      description: Confirm the payment of a reservation, updating payment status,
        amount, date, and reservation status
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Reservation payment details (payment_amount)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Reservation'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Confirm payment for a reservation
      tags:
      - Reservations
  /reservations/{id}/return-keys:
    put:
      description: Marks the keys as returned for a reservation and updates its status
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Mark reservation keys as returned
      tags:
      - Reservations
  /reservations/{id}/take-keys:
    put:
      description: Marks the keys as taken for a confirmed reservation and updates
        its status
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Mark reservation keys as taken
      tags:
      - Reservations
  /reservations/date-range:
    get:
      description: 'Retrieve all reservations between start_date and end_date (format:
        yyyy-MM-dd)'
      parameters:
      - description: Start date in format yyyy-MM-dd
        in: query
        name: start_date
        required: true
        type: string
      - description: End date in format yyyy-MM-dd
        in: query
        name: end_date
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Reservation'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get reservations by date range
      tags:
      - Reservations
  /reservations/import-salon:
    post:
      description: Imports reservations data from a CSV file for salon spaces
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Import salon reservations from CSV
      tags:
      - Reservations
...
  /reservations/resident/{residentId}:
    get:
      description: Retrieve all reservations associated with a specific resident
      parameters:
      - description: Resident ID
        in: path
        name: residentId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Reservation'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get reservations by resident ID
      tags:
      - Reservations
  /reservations/space/{space}:
    get:
      description: Retrieve all reservations for a specific space
      parameters:
      - description: Space type
        in: path
        name: space
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Reservation'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get reservations by space type
      tags:
      - Reservations
...
  /reservations/status/{status}:
    get:
      description: Retrieve all reservations filtered by their status
      parameters:
      - description: Reservation status
        in: path
        name: status
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Reservation'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get reservations by status
      tags:
      - Reservations
  /reservations/upcoming:
    get:
      description: Retrieve all upcoming reservations
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Reservation'
            type: array
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get upcoming reservations
      tags:
      - Reservations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"